split preprocessor implementation - i thk we will move it to .NET Core plugin for other languages like (F#)
